src/hooks/common/use-keyboard-directions/use-keyboard-directions.tsx
import { useState, useEffect } from "react";
import { Direction } from "./types";

function useKeyboardDirections() {
  const [direction, setDirection] = useState<Direction>("IDLE");

  useEffect(() => {
    const handleKeyDown = (event: KeyboardEvent) => {
      switch (event.key) {
        case "ArrowUp":
          setDirection("UP");
          break;
        case "ArrowDown":
          setDirection("DOWN");
          break;
        case "ArrowLeft":
          setDirection("LEFT");
          break;
        case "ArrowRight":
          setDirection("RIGHT");
          break;
        default:
          break;
      }
    };

    window.addEventListener("keydown", handleKeyDown);

    return () => {
      window.removeEventListener("keydown", handleKeyDown);
    };
  }, []);

  return direction;
}

export default useKeyboardDirections;
